Renamed Results.vue to ResultsContainer.vue
To comply with `vue/multi-word-component-names` rule
